Réalisation des entités

+++ src/main/java/com/terroir/repositories/OrigineRepo.java
@@ -3,5 +3,5 @@ package com.terroir.repositories;
 import com.terroir.entities.Origine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrigineRepo extends JpaRepository<Origine,Long> {
+public interface OrigineRepo extends JpaRepository<Origine,Integer> {
 }
